+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/delete.go
@@ -4,7 +4,9 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
 )
@@ -19,8 +21,17 @@ func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
 
 		name, namespace := key.RemoteWriteAPIEndpointConfigSecretNameAndNamespace(cluster, r.Installation, r.Provider)
 
-		current, err := r.k8sClient.K8sClient().CoreV1().Secrets(namespace).Get(ctx, name, metav1.GetOptions{})
+		_, err = r.k8sClient.K8sClient().CoreV1().Secrets(namespace).Get(ctx, name, metav1.GetOptions{})
+		if apierrors.IsNotFound(err) {
+			// We ignore cases where the secret is not found (it it was manually deleted for instance)
+			return nil
+		} else if err != nil {
+			return microerror.Mask(err)
+		}
 
+		// Delete the finalizer
+		patch := []byte(`{"metadata":{"finalizers":null}}`)
+		current, err := r.k8sClient.K8sClient().CoreV1().Secrets(namespace).Patch(ctx, name, types.StrategicMergePatchType, patch, metav1.PatchOptions{})
 		if err != nil {
 			return microerror.Mask(err)
 		}