+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add finalizer to remote-write-config to block cluster deletion until PMO deleted the secret.
+
 ## [4.35.0] - 2023-04-11
 
 ### Changed